requirements.txt
# ── Core Framework ──
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── JWT / Crypto ──
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0

# ── Redis ──
redis>=5.0.0
fakeredis>=2.21.0

# ── HTTP (JWKS fetching) ──
httpx>=0.27.0

# ── Testing ──
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── Logging ──
structlog>=24.1.0
